Add more gpt_models

+++ drdictaphone/model/gpt_model.py
@@ -9,8 +9,8 @@ class GptModel(BaseModel):
   model_config = ConfigDict(extra = 'forbid')
 
   name: str
-  input_cost: int
-  output_cost: int
+  input_cost: float
+  output_cost: float
 
   def __str__(self):
     return f'{self.name}, {self.input_cost}, {self.output_cost}'
